+++ src/main.tsx
@@ -1,14 +1,14 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './assets/styles/global.css'
-import App from './App'
-import { store } from './app/store'
-import { Provider } from 'react-redux'
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import './assets/styles/global.css';
+import App from './App';
+import { store } from './redux/store';
+import { Provider } from 'react-redux';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <Provider store={store}>
-    <App />
+      <App />
     </Provider>
-  </StrictMode>,
-)
+  </StrictMode>
+);
