+++ src/types/contactTypes.ts
@@ -1,0 +1,7 @@
+
+export interface ContactCard {
+	id: string;
+	firstName: string;
+	lastName: string;
+  }
+
